Complete indexed sprite plan: encoding, migration, and flash savings
Extract shared sprite_encoding module used by sprite_converter.py and
png_to_spritesheet.py (--indexed auto|4|8|16, --quantize, --max-colors).

Add reindex_header.py to convert legacy RGB565 headers without PNG sources,
compare_indexed_savings.py for flash estimates, and indexed background rendering
via setBackgroundAsset/drawBackgroundAsset.

Migrate sprite_totoro_baby, sprite_acorn, and image_acorn_catch_bg to 8-bit
SpriteAsset format; update Pet Totoro and Acorn Catch scenes. Firmware flash
drops from ~951 KB to ~866 KB (~85 KB saved).

+++ GameBase/src/SpriteAsset.h
@@ -21,7 +21,7 @@ struct SpriteAsset {
   uint8_t bpp;
   uint16_t sheetWidth;
   uint16_t sheetHeight;
-  uint8_t paletteCount;
+  uint16_t paletteCount;
   const uint16_t *palette;
   const void *pixels;
   const uint8_t *mask;
